python-server/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv